+++ src/cli/console.ts
@@ -1,15 +1,19 @@
 import chalk from "chalk";
 
-export function log(msg: any): void {
-  // tslint:disable-next-line: no-console
-  console.log(msg);
+function create(name: Extract<keyof typeof console, "log" | "error">) {
+  const method = console[name];
+  return (msg: any) => (
+    method(chalk.cyan(msg))
+  );
 }
+
+export const log = create("log");
+export const error = create("error");
 
 export function logError(err: Error) {
   if (err.stack) {
-    // tslint:disable-next-line: no-console
-    console.error(err.stack);
+    error(err.stack);
   }
-  // tslint:disable-next-line: no-console
-  console.error(`${chalk.bgRed.bold(" error ")} ${err.message || err}`);
+  const tag = chalk.bold.bgWhite.red.inverse(" error ");
+  error(`${tag} ${err.message || err}`);
 }